==> Paper/Regression/1_preprocessing.Rmd <==
---
title: 'Author perspective'
author: "Kevin Wirtz (kevin.wirtz@unistra.fr)"
date: "Updated `r format(Sys.time(), '%B %d, %Y')`"
output:
  html_notebook:
    code_folding: hide
    df_print: paged
    toc: true
    toc_depth: 2
    toc_float:
      collapsed: false
    theme: flatly
---

```{r setup, include=FALSE}

rm(list = ls(all.names = TRUE)) #will clear all objects includes hidden objects.
gc() #free up memrory and report the memory usage.
graphics.off()

Sys.setenv(LANG = "en") # For english language
options(scipen = 5) # To deactivate annoying scientific number notation

### Load packages
library(tidyverse) # Collection of all the good stuff like dplyr, ggplot2 ect.
library(magrittr) # For extra-piping operators (eg. %<>%)
library(glue)
library(data.table)
library(dplyr)
library(ks)
# Descriptives
#library(skimr)
library(stargazer)
library(corrplot)
# ML
library(xgboost)
library(caret)  
library(caTools)
library(e1071)



```

Import data

```{r} 

df <- read_csv("../../Data/regression.csv")
df$year <- as.factor(df$year)

```
Init name of variables of interest

```{r} 
novelty_cat <- c("Technical Advance","Interesting Hypothesis","Novel Drug Target")
variables <- c("nb_ref", "nb_meshterms", "year", "nb_aut", "journal_SJR", "is_review", "journal_age","sum_deg_cen_cumsum","sum_deg_cen_cumsum_mean", "journal_main_cat")
indicators <- c("uzzi_ref","uzzi_mesh","lee_ref","lee_mesh","foster_ref","foster_mesh",
                "wang_ref","wang_mesh","shibayama_abstract","shibayama_title")
indicators_fw <- paste0(indicators, "_fw")
independant <-  c("author_intra_abstract","author_inter_abstract", "author_intra_title", "author_inter_title")
independant_fw <- paste0(independant, "_fw")
impact <-  c("nb_cit","DI1","DI5","DI1nok","DeIn","Breadth","Depth")
impact_fw <- paste0(impact, "_fw")
```


Clean journal age

```{r, eval = FALSE} 
df$journal_age <- replace(df$journal_age, which(df$journal_age < 0), NA)
```

Create novelty binary
```{r} 
df <- df %>% mutate(novel_f1000 = ifelse(main_category %in% novelty_cat, 1, 0))
```

```{r} 
df %<>%
  mutate(sum_deg_cen_cumsum_mean = sum_deg_cen_cumsum/nb_aut,
         uzzi_ref = -uzzi_ref,
         uzzi_mesh = -uzzi_mesh)
```

replace na where it make sense


```{r} 
df %<>%
  filter(nb_ref>1 & nb_meshterms>1 & nb_aut>1 & journal_ISSN != "" & journal_main_cat != "") %>%
  drop_na(journal_ISSN)
```


```{r, eval = FALSE} 
df %<>%
  replace_na(list(nb_ref = 0, nb_meshterms = 0))
```

Field weight by journal category from scimago
```{r} 
df %<>%
  group_by(journal_main_cat, year) %>%
  mutate(nb_cit_fw = nb_cit %>% percent_rank(),
         wang_mesh_fw = wang_mesh %>% percent_rank(),
         wang_ref_fw = wang_ref %>% percent_rank(),
         uzzi_mesh_fw = uzzi_mesh %>% percent_rank(),
         uzzi_ref_fw = uzzi_ref %>% percent_rank(),
         foster_mesh_fw = foster_mesh %>% percent_rank(),
         foster_ref_fw = foster_ref %>% percent_rank(),
         lee_mesh_fw = lee_mesh %>% percent_rank(),
         lee_ref_fw = lee_ref %>% percent_rank(),
         shibayama_abstract_fw = shibayama_abstract %>% percent_rank(),
         shibayama_title_fw = shibayama_title %>% percent_rank(),
         author_intra_abstract_fw = author_intra_abstract %>% percent_rank(),
         author_inter_abstract_fw = author_inter_abstract %>% percent_rank(),
         author_intra_title_fw = author_intra_title %>% percent_rank(),
         author_inter_title_fw = author_inter_title %>% percent_rank(),
         DI5_fw = DI5 %>% percent_rank(),
         DI1_fw = DI1 %>% percent_rank(),
         DI1nok_fw = DI1nok %>% percent_rank(),
         DeIn_fw = DeIn %>% percent_rank(),
         Breadth_fw = Breadth %>% percent_rank(),
         Depth_fw = Depth %>% percent_rank(),
           ) %>%
  ungroup()

fwrite(df,'reg_df.csv')
```

Missing
```{r} 
df %>% naniar::gg_miss_var()
```

```{r} 
summary_col <- function(col) {
   print(sum(!is.na(df[[col]])))
   print(summary(df[[col]]))
   print(sd(df[[col]],na.rm=TRUE))
}
summary_col(col="uzzi_ref")
```

citation normal vs f1000
```{r} 
df_temp = df %>% 
  drop_na(main_category)
mean(df_temp$nb_cit, na.rm = TRUE)

df_temp = df[is.na(df$main_category),]
mean(df_temp$nb_cit, na.rm = TRUE)

```

For regression only keep paper matched from f1000
```{r} 
df_f1000 <- df %>% 
  drop_na(main_category)
```

# Stats ####

```{r} 
df_f1000 %>% count(year) %>%
  ggplot(aes(x = year, y = n)) + 
  geom_col() +
  theme_classic() + 
  ggtitle("Number of papers matched with F1000 2000-2005") +
  theme(plot.title = element_text(hjust = 0.5))
```
```{r} 

f1000_cat = c()
f1000_cat_binary = c()

for (i in novelty_cat){
  name_cat = str_replace_all(i, c(" " = "_","/"="_",":"="_"))
  name_cat_binary = str_replace_all(paste(i,"_binary"), c(" " = "_","/"="_",":"="_"))
  df_f1000[name_cat] <- lengths(regmatches(df_f1000$categories, gregexpr(i,df_f1000$categories)))
  df_f1000[name_cat_binary] <- ifelse(df_f1000[name_cat] >= 1, 1, 0)
  f1000_cat <- append(f1000_cat,name_cat)
  f1000_cat_binary <- append(f1000_cat_binary,name_cat_binary)
}
```

```{r} 
df_f1000 <- df_f1000 %>% 
  mutate(novel_f1000_v2 = case_when(Technical_Advance__binary == 1 ~ 1,
                                    Interesting_Hypothesis__binary == 1 ~ 1,
                                    Novel_Drug_Target__binary == 1 ~ 1)) 
df_f1000$novel_f1000_v2 <- replace(df_f1000$novel_f1000_v2, is.na(df_f1000$novel_f1000_v2), 0)
```

Data for reg1 on recombinatorial novelty
```{r} 
df %>%
  select( all_of(c(indicators_fw,variables,independant_fw)) )
```

```{r} 
fwrite(df,'../../Data/reg_df.csv')
```
