==> Paper/Regression/3_regression_3.Rmd <==
---
title: 'Author perspective'
author: "Kevin Wirtz (kevin.wirtz@unistra.fr)"
date: "Updated `r format(Sys.time(), '%B %d, %Y')`"
output:
  html_notebook:
    code_folding: hide
    df_print: paged
    toc: true
    toc_depth: 2
    toc_float:
      collapsed: false
    theme: flatly
---

```{r setup, include=FALSE}

rm(list = ls(all.names = TRUE)) #will clear all objects includes hidden objects.
gc() #free up memrory and report the memory usage.
graphics.off()

Sys.setenv(LANG = "en") # For english language
options(scipen = 5) # To deactivate annoying scientific number notation

### Load packages
library(tidyverse) # Collection of all the good stuff like dplyr, ggplot2 ect.
library(magrittr) # For extra-piping operators (eg. %<>%)
library(glue)
# Descriptives
#library(skimr)
library(stargazer)
library(corrplot)
# ML
library(xgboost)
library(caret)  
library(caTools)
library(e1071)
```

Link between authors and impact measures

Import data
```{r} 
df = read_rds(file = "../../Data/regression_3.rds")
```


define dependant variables
```{r} 
independant <-  c("author_intra_abstract","author_inter_abstract", "author_intra_title", "author_inter_title")
independant_fw <- paste0(independant, "_fw")
variables <- c("nb_ref", "nb_meshterms", "year", "nb_aut", "journal_SJR", "is_review", "journal_age","sum_deg_cen_cumsum")
impact <-  c("nb_cit","DI1","DI5","DI1nok","DeIn","Breadth","Depth")
impact_fw <- paste0(impact, "_fw")
```

################################################################################
####################### Abstract on impact #######################################
################################################################################

```{r} 


mod_1 = lm(nb_cit_fw ~ author_intra_abstract_fw + author_inter_abstract_fw+
           nb_ref + nb_meshterms + year + nb_aut + journal_SJR + is_review + journal_age + sum_deg_cen_cumsum,
                                              data = df)

mod_2 = lm(DI1nok_fw ~ author_intra_abstract_fw + author_inter_abstract_fw+
           nb_ref + nb_meshterms + year + nb_aut + journal_SJR + is_review + journal_age + sum_deg_cen_cumsum,
                                              data = df)

mod_3 = lm(Depth_fw ~ author_intra_abstract_fw + author_inter_abstract_fw+
           nb_ref + nb_meshterms + year + nb_aut + journal_SJR + is_review + journal_age + sum_deg_cen_cumsum,
                                              data = df)

stargazer(mod_1, mod_2, mod_3,
          type = 'text', omit = variables)

```

################################################################################
####################### Title on impact #######################################
################################################################################

```{r} 


mod_1 = lm(nb_cit_fw ~ author_intra_title_fw + author_inter_title_fw+
           nb_ref + nb_meshterms + year + nb_aut + journal_SJR + is_review + journal_age + sum_deg_cen_cumsum,
                                              data = df)

mod_2 = lm(DI1nok_fw ~ author_intra_title_fw + author_inter_title_fw+
           nb_ref + nb_meshterms + year + nb_aut + journal_SJR + is_review + journal_age + sum_deg_cen_cumsum,
                                              data = df)

mod_3 = lm(Depth_fw ~ author_intra_title_fw + author_inter_title_fw+
           nb_ref + nb_meshterms + year + nb_aut + journal_SJR + is_review + journal_age + sum_deg_cen_cumsum,
                                              data = df)

stargazer(mod_1, mod_2, mod_3,
          type = 'text', omit = variables)

```

################################################################################
####################### Abstract on impact NO CV#######################################
################################################################################

```{r} 


mod_1 = lm(nb_cit_fw ~ author_intra_abstract_fw + author_inter_abstract_fw,
                                              data = df)

mod_2 = lm(DI1nok_fw ~ author_intra_abstract_fw + author_inter_abstract_fw,
                                              data = df)

mod_3 = lm(Depth_fw ~ author_intra_abstract_fw + author_inter_abstract_fw,
                                              data = df)

stargazer(mod_1, mod_2, mod_3,
          type = 'text', omit = variables)

```


################################################################################
####################### Title on impact NO CV#######################################
################################################################################

```{r} 


mod_1 = lm(nb_cit_fw ~ author_intra_title_fw + author_inter_title_fw,
                                              data = df)

mod_2 = lm(DI1nok_fw ~ author_intra_title_fw + author_inter_title_fw,
                                              data = df)

mod_3 = lm(Depth_fw ~ author_intra_title_fw + author_inter_title_fw
                                              data = df)

stargazer(mod_1, mod_2, mod_3,
          type = 'text', omit = variables)

```
